Added DebugTools.java, more work on Player Class, Added Some testing components
DebugTools will be a collection of functions to speed up testing, right now
there is one function to simulate input from a user, for example when we are
entering a move. I started some of the tests, but they are very basic due to thelimited functionality of the Player right now.