Starting work on the network overhaul
I have started QuoridorClient.java and am working on implementing the
functionality described in the log in the actual .java file of this object.
Each player will have an instance of the client which has an input stream
that will eventually get its input from the GUI for moves/wall placements.
The client will continue running and buffering the moves input and when
it is the player's turn a method will be called to read the move/placement.