Working on making connections between server(s) & client
The Quoridor client will now take in CLA's and can parse them, it will be used
to make connections and store the socket information so the client can continue
to ask for moves, walls, booting, etc.. This will eventually start up a new
board and take care of all the players input.

I also modified the .gitignore file so that you don't have to worry about
adding files with a '~' at the end of them anymore. Also working on some tests.
More will be possible by using netcat.

The client should be mostly done (not starting a new game) by tommorow.