Added an interface to use for the Client/Server message passing scheme.
The client can now connect to sockets that are open and waiting for someone
to connect to them.

The interface (Messages.java) contains constant strings that will be used
between the Client/Server especcially when writing AI, and when making a
parser for the movement of pawns and placement of walls.