DO NOT PULL BEFORE DEMO. BROKEN UI.
I revived the Board class by adding a Player array and Wall array
and methods to move players and add walls to the board.

Wrote tests for all methods. Will need to add more conditions to
move validation as we implement AI.

The UI is currently broken since I started moving the logic in the
UI classes to separate classes but did not have time to get the UI
to an error-free state. We'll talk about this at our meeting after class today.