Walls can now be placed via terminal move entries.
As I move forward with implementing logic from the Controller class,
I need to keep an eye on the position coordinates, as they seem to
be reversed in some places. Also, I need to update the wall class
because I have walls starting at the given coordinates, not centered
on the given coordinates. Easy fix though.