Still a couple occasional kinks in the move validation and wall placement because when I wrote the wall class I thought the (x,y) coordinates would be at the top/left of the wall rather than the middle. Not a bad fix, I just need to take a break from this code and work on another project. I'll iron out the kinks and write tests in the next few days.
I started adding Javadoc comments to the code to build an API to make the
code easier to follow. You can view the API in Eclipse or a browser just
like the Java API.

Also, you can run Game.java with the command line args "2 simpleGame.txt"
and watch a game play. Feel free to create a more complex file to demo
if we don't have the AI running by the next class presentation.