Added a lot of the MoveServer implementation.
The MoveServer starts up and listens on a port for a connection. This connection
should hopefully come from a GameClient. When the connection is made, a Scanner
and PrintStream open and when fed input the server has a chance to reply.

I am working on handling the different input from the gameClient now.