Added further message protocols.
One issue I am having is with entering moves. I don't know what I am supposed to be using to get the input from the user. Maybe it will come to me later, if
anyone has an idea? The issue is when it goes to get the move it hasn't been
typed yet and throws an exception because there is no input there.