Made booting work over the networking interface.
When an illegal move is attempted the Client looks up who tried to move last and alerts them that they made an invalid move and they are given a message, as
per the protocol, and everyone else gets the message as well. The moveserver
is then exited from and the game continues where it left off.