Starting work on Artie, our Resident AI
This AI will extend our MoveServer to save some of that functionality and make
it quicker/easier to debug as the functionality of the MoveServer works well.
All that needs to be changed are the methods to return moves, which will
rely on how I weight moves/walls, and how I create this move tree.

Reminder.. Our agreed upon AI_Identifier is "tactical"